+++ clang/lib/AST/ExprConstant.cpp
@@ -19740,6 +19740,8 @@ bool IntExprEvaluator::VisitBinaryOperator(const BinaryOperator *E) {
 bool IntExprEvaluator::VisitUnaryExprOrTypeTraitExpr(
                                     const UnaryExprOrTypeTraitExpr *E) {
   switch(E->getKind()) {
+  case UETT_AddrSpaceOf:
+    return Success(E->getAddressSpaceQueryResult(Info.Ctx), E);
   case UETT_PreferredAlignOf:
   case UETT_AlignOf: {
     if (E->isArgumentType())