+++ clang/lib/AST/StmtPrinter.cpp
@@ -1717,6 +1717,10 @@ void StmtPrinter::VisitUnaryExprOrTypeTraitExpr(
     OS << '(';
     Node->getArgumentType().print(OS, Policy);
     OS << ')';
+  } else if (Node->getKind() == UETT_AddrSpaceOf) {
+    OS << '(';
+    PrintExpr(Node->getArgumentExpr());
+    OS << ')';
   } else {
     OS << " ";
     PrintExpr(Node->getArgumentExpr());